requirements.txt
Django
djangorestframework
django-redis-cache